+++ rdservice/rdservice.cpp
@@ -84,14 +84,6 @@ MainObject::MainObject(QObject *parent)
   }
 
   //
-  // Ensure that we are 'root'
-  //
-  if(geteuid()!=0) {
-    rda->syslog(LOG_ERR,"this service requires root");
-    exit(RDApplication::ExitNoPerms);
-  }
-
-  //
   // Process Startup Options
   //
   for(unsigned i=0;i<rda->cmdSwitch()->keys();i++) {